# montessori-match-1-swagger.yaml
swagger: '2.0'
info:
  version: '1'
  title: 'MM'
  description: 'first iteration'
securityDefinitions:
  JWT:
    description: JWT token containing the user's ID in the body.
    type: apiKey
    name: authToken
    in: header
paths:
  /auth:
    post:
      description: Creates a user in the system. Returns an auth token.
      tags:
      - auth
      parameters: 
      - name: credentials
        in: body
        required: true
        schema:
          type: object
          properties:
            email:
              type: string
              format: email
            password:
              type: string
      responses:
        201:
          description: success
          schema:
            type: object
            properties:
              token:
                type: string
        422: 
          $ref: '#/responses/422'
    get:
      description: Logs a user in. Returns an auth token.
      tags:
      - auth
      parameters:
      - name: email
        in: query
        required: true
        type: string
      - name: password
        in: query
        required: true
        type: string
      responses:
        200:
          description: success
          schema:
            type: object
            properties:
              token:
                type: string
        422: 
          $ref: '#/responses/422'
...
host: virtserver.swaggerhub.com
basePath: /dsudia/montessori-match/1
schemes:
- https
consumes:
- application/json
produces:
- application/json
definitions:
...
  Error:
    type: object
    properties:
      code:
        type: number
      message:
        type: string
...
responses:
...
  422:
    description: bad parameters
    schema:
      $ref: '#/definitions/Error' 
...